requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# API integrations
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.107.0
PyGithub==2.1.1

# OpenAI integration
openai==1.3.5

# Utilities
requests==2.31.0
aiohttp==3.8.6
python-multipart==0.0.6 